# apps/web/public/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Rate Card Benchmarking API
  description: |
    Comprehensive REST API for Rate Card Benchmarking Engine with advanced analytics,
    predictive forecasting, AI insights, and supplier intelligence.
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging-api.example.com/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
security:
  - BearerAuth: []
...
paths:
...
  /rate-cards/clusters:
    get:
      tags:
        - Clustering
      summary: List clusters
      parameters:
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Clusters retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cluster'
  /rate-cards/suppliers/{id}/intelligence:
    get:
      tags:
        - Suppliers
      summary: Get supplier intelligence
      description: Retrieve comprehensive supplier intelligence and scoring
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Supplier intelligence retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierIntelligence'
...
  /rate-cards/segments:
    get:
      tags:
        - Segments
      summary: List saved segments
      parameters:
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Segments retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Segment'
    
    post:
      tags:
        - Segments
      summary: Create segment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SegmentCreate'
      responses:
        '201':
          description: Segment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Segment'
  /rate-cards/reports/schedule:
    post:
      tags:
        - Reports
      summary: Schedule report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduledReportCreate'
      responses:
        '201':
          description: Report scheduled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledReport'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Cluster:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        memberCount:
          type: integer
        avgRate:
          type: number
        consolidationSavings:
          type: number
        characteristics:
          type: object
    SupplierIntelligence:
      type: object
      properties:
        supplierId:
          type: string
        overallScore:
          type: number
        priceCompetitiveness:
          type: number
        geographicCoverage:
          type: number
        rateStability:
          type: number
        ranking:
          type: integer
        trend:
          type: string
...
    Segment:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        filters:
          type: object
        shared:
          type: boolean
    SegmentCreate:
      type: object
      required:
        - tenantId
        - name
        - filters
      properties:
        tenantId:
          type: string
        name:
          type: string
        description:
          type: string
        filters:
          type: object
    ScheduledReport:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        frequency:
          type: string
        recipients:
          type: array
          items:
            type: string
        nextRun:
          type: string
          format: date-time
    ScheduledReportCreate:
      type: object
      required:
        - tenantId
        - name
        - type
        - frequency
        - recipients
      properties:
        tenantId:
          type: string
        name:
          type: string
        type:
          type: string
        frequency:
          type: string
        recipients:
          type: array
          items:
            type: string
...